# perf(api): optimize API_GetGameList