Resolves https://retroachievementsorg.sentry.io/issues/45318267/?environment=production&query=is%3Aunresolved&referrer=issue-stream.

This PR significantly optimizes _API_GetGameList.php_. There is still a slight bottleneck for the server to actually build the gigantic payload, but the internal code is now fairly well optimized.

With 10 iterations for the internal code with `i` set to 21 and `h` set to 1:
* Base query: went from ~850ms to ~24ms.
* Hash lookup: went from ~15,000ms to ~92ms.

Data is now pre-aggregated in subqueries before joining it with the base query. Nested loops for hash lookups have been replaced with indexed lookups.